Use memcpy for unaligned access
This is the most standard, safe and portable way. But performance
suffers slightly - up to 4% on ARM. If a target doesn't support
unaligned access, performance can degrade considerably. But such
platforms are becoming rare these days. The previous approach was not
standard, and modern compilers could do nasty things with it.

Now -mno-unaligned-access compiler option should be used to force
disabling unaligned access if it is disabled on target at runtime.